requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=0.6,!=0.7,<1.0

jinja2
oslo.config>=1.6.0  # Apache-2.0
oslo.i18n>=1.3.0  # Apache-2.0
oslo.serialization>=1.2.0               # Apache-2.0
oslo.utils>=1.2.0                       # Apache-2.0
python-keystoneclient>=1.0.0
python-novaclient>=2.18.0
python-heatclient>=0.2.9
PyYAML>=3.1.0
